Fixed callAPIService incompatibility with Railo
When deserializing JSON, Railo behaves differently from Coldfusion if presented with the string 'null'. In that case, Railo completely deletes the variable, making it unavailable, and causing an error in callAPIService - whereas Coldfusion returns the string null. The proposed fix should fix the issue without sideffects. This presents consistently when using realtime subscriptions.